AWS_MWAA/requirements.txt
apache-airflow==2.6.3
boto3==1.28.0
scikit-learn==1.3.0
joblib==1.3.1
